Add the ability to add headers to web client requests
Add an option named headers that users can use to add headers to client requests
the web transport makes. You'd use this by adding the object:

```javascript
{
  web: {
    headers: {
      'client-id': 'test-client'
    }
  }
}
```

to the use statement for the transport. This adds the header 'client-id' with
the value 'test-client' to requests that the web client makes to other
seneca services.